+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [3.0.9-beta] — 2026-06-TBD
+- Additional CSS field added by WP is breaking out of our tabs - FIXED
+- Bug with some select items rendering as an array in JSON instead of object - FIXED
+
 ## [3.0.8-beta] — 2026-05-18
 - font_italic field removed from add_typography_group() as it is part for font_weight field - CHANGED
 - added helper function ayecode_get_current_url( $with_query_string = true ) - ADDED